+++ webknossos/webknossos/annotation/annotation.py
@@ -80,7 +80,12 @@ from ..dataset_properties import (
 from ..geometry import NDBoundingBox, Vec3Int
 from ..proofreading.agglomerate_graph_data import AgglomerateGraphData
 from ..skeleton import Skeleton
-from ..utils import get_executor_for_args, is_fs_path, time_since_epoch_in_ms
+from ..utils import (
+    get_executor_for_args,
+    is_fs_path,
+    time_since_epoch_in_ms,
+    warn_deprecated,
+)
 from ._nml_conversion import annotation_to_nml, nml_to_skeleton
 from .volume_layer import SegmentInformation, VolumeLayer
 
@@ -1534,8 +1539,12 @@ class RemoteAnnotation(Annotation):
         return file_path
 
     def get_agglomerate_graph(self, agglomerate_id: int) -> AgglomerateGraphData:
+        warn_deprecated("get_agglomerate_graph", "get_agglomerate_graph_data")
+        return self.get_agglomerate_graph_data(agglomerate_id)
+
+    def get_agglomerate_graph_data(self, agglomerate_id: int) -> AgglomerateGraphData:
         """
-        Get the agglomerate graph for the specified agglomerate id.
+        Get the agglomerate graph data for the specified agglomerate id.
         This works only for proofreading annotations that have only a single volume layer.
 
         Args: